fix: map component added
