+++ src/index.css
@@ -1,0 +1,13 @@
+.App {
+    display: flex;
+    min-height: 100vh;
+    justify-items: center;
+    justify-content: center;
+    align-items: center;
+    background-color: rgb(135, 227, 255);
+}
+
+* {
+    margin: 0;
+    padding: 0;
+}
